fix: increase the number of items fetched per page in the cursor to 20
There's a slight problem with the cursor used for row sampling where
each connection can only open one cursor at once. So if there are 10+
tables all with a lot of data, it can cause a deadlock trying to process
the 11th table. This isn't a proper fix but coupled with the increased
connection cap it should help some cursors close before others are
opened